Add BaseApplication and BaseWorker. Separate logic between classes
