+++ pymongo_change_stream_reader/change_stream_reading/__init__.py
@@ -1,2 +1,8 @@
-from .builder import build_change_stream_reader_process
+from .builder import (
+    build_change_stream_reader_process,
+    build_change_stream_reader_worker,
+)
 from .change_stream_reader import ChangeStreamReader
+from .change_handler import ChangeHandler
+from .resume_token import RetrieveResumeToken
+from .watch import ChangeStreamWatch
