+++ pymongo_change_stream_reader/change_stream_reading/builder.py
@@ -1,10 +1,22 @@
-from asyncio import Queue
 from multiprocessing import Process
+from multiprocessing import Queue
 
-from .change_stream_reader import ChangeStreamReader
+from bson.raw_bson import RawBSONDocument
+from pymongo import MongoClient
+
+from pymongo_change_stream_reader.app_context import ApplicationContext
+from pymongo_change_stream_reader.base_worker import BaseWorker
+from pymongo_change_stream_reader.models import ProcessData
+from pymongo_change_stream_reader.settings import (
+    FullDocumentBeforeChange,
+    FullDocument,
+)
 from pymongo_change_stream_reader.settings import Settings
 from pymongo_change_stream_reader.utils import TaskIdGenerator
-from pymongo_change_stream_reader.models import ProcessData
+from .change_handler import ChangeHandler
+from .change_stream_reader import ChangeStreamReader
+from .resume_token import RetrieveResumeToken
+from .watch import ChangeStreamWatch
 
 
 def build_change_stream_reader_process(
@@ -41,10 +53,78 @@ def build_change_stream_reader_process(
         'queue_get_timeout': settings.queue_get_timeout,
         'queue_put_timeout': settings.queue_put_timeout,
     }
-    process = Process(target=run_change_stream_reader, kwargs=kwargs)
+    process = Process(target=ChangeStreamReaderContext.run_application, kwargs=kwargs)
     return ProcessData(task_id=task_id, process=process, kwargs=kwargs)
 
 
-def run_change_stream_reader(**kwargs):
-    worker = ChangeStreamReader(**kwargs)
-    worker.run()
+def build_change_stream_reader_worker(
+    manager_pid: int,
+    manager_create_time: float,
+    task_id: int,
+    producer_queues: dict[int, Queue],
+    request_queue: Queue,
+    response_queue: Queue,
+    committer_queue: Queue,
+    mongo_uri: str,
+    token_mongo_uri: str,
+    token_database: str,
+    token_collection: str,
+    stream_reader_name: str,
+    database: str | None = None,
+    collection: str | None = None,
+    pipeline: list[dict] | None = None,
+    full_document_before_change: FullDocumentBeforeChange = (
+            FullDocumentBeforeChange.when_available
+    ),
+    full_document: FullDocument = FullDocument.when_available,
+    reader_batch_size: int | None = None,
+    queue_get_timeout: int = 1,
+    queue_put_timeout: int = 10,
+) -> BaseWorker:
+    token_mongo_client = MongoClient(
+        host=token_mongo_uri,
+    )
+    mongo_client = MongoClient(
+        host=mongo_uri,
+        document_class=RawBSONDocument
+    )
+    token_retriever = RetrieveResumeToken(
+        stream_reader_name=stream_reader_name,
+        token_mongo_client=token_mongo_client,
+        token_database=token_database,
+        token_collection=token_collection,
+    )
+    watcher = ChangeStreamWatch(
+        mongo_client=mongo_client,
+        collection=collection,
+        database=database,
+        pipeline=pipeline,
+        full_document_before_change=full_document_before_change,
+        full_document=full_document,
+        reader_batch_size=reader_batch_size,
+    )
+    change_handler = ChangeHandler(
+        committer_queue=committer_queue,
+        producer_queues=producer_queues,
+        queue_put_timeout=queue_put_timeout,
+    )
+    application = ChangeStreamReader(
+        token_retriever=token_retriever,
+        watcher=watcher,
+        change_handler=change_handler,
+    )
+    worker = BaseWorker(
+        manager_pid=manager_pid,
+        manager_create_time=manager_create_time,
+        task_id=task_id,
+        request_queue=request_queue,
+        response_queue=response_queue,
+        queue_get_timeout=queue_get_timeout,
+        queue_put_timeout=queue_put_timeout,
+        application=application,
+    )
+    return worker
+
+
+class ChangeStreamReaderContext(ApplicationContext):
+    build_worker = staticmethod(build_change_stream_reader_worker)
