+++ pymongo_change_stream_reader/producing/__init__.py
@@ -1,2 +1,4 @@
-from .builder import build_producer_process
-from .producer import ProducerFlow
+from .builder import build_producer_process, build_producer_worker
+from .producer_flow import ProducerFlow
+from .producer import Producer
+from .change_event_handler import ChangeEventHandler
