+++ pymongo_change_stream_reader/producing/builder.py
@@ -1,10 +1,18 @@
 from asyncio import Queue
 from multiprocessing import Process
+from typing import Any
 
-from .producer import ProducerFlow
+from confluent_kafka import Producer as KafkaProducer
+from confluent_kafka.admin import AdminClient
+
+from pymongo_change_stream_reader.app_context import ApplicationContext
+from pymongo_change_stream_reader.base_worker import BaseWorker
+from pymongo_change_stream_reader.models import ProcessData
 from pymongo_change_stream_reader.settings import Settings, NewTopicConfiguration
 from pymongo_change_stream_reader.utils import TaskIdGenerator
-from pymongo_change_stream_reader.models import ProcessData
+from .change_event_handler import ChangeEventHandler
+from .producer import Producer
+from .producer_flow import ProducerFlow
 
 
 def build_producer_process(
@@ -36,7 +44,7 @@ def build_producer_process(
         'queue_get_timeout': settings.queue_get_timeout,
         'queue_put_timeout': settings.queue_put_timeout,
     }
-    process = Process(target=run_producer, kwargs=kwargs)
+    process = Process(target=ProducerFlowContext.run_application, kwargs=kwargs)
     return ProcessData(
         task_id=task_id,
         process=process,
@@ -44,6 +52,63 @@ def build_producer_process(
     )
 
 
-def run_producer(**kwargs):
-    worker = ProducerFlow(**kwargs)
-    worker.run()
+def build_producer_worker(
+    manager_pid: int,
+    manager_create_time: float,
+    task_id: int,
+    producer_queue: Queue,
+    request_queue: Queue,
+    response_queue: Queue,
+    committer_queue: Queue,
+    stream_reader_name: str,
+    kafka_bootstrap_servers: str,
+    new_topic_configuration: dict[str, Any],
+    kafka_prefix: str,
+    kafka_producer_config: dict[str, str],
+    queue_get_timeout: int,
+    queue_put_timeout: int,
+) -> BaseWorker:
+    new_topic_configu = NewTopicConfiguration.parse_obj(new_topic_configuration)
+
+    kafka_config = {}
+    kafka_config.update(kafka_producer_config)
+    kafka_config.update(
+        {
+            'bootstrap.servers': kafka_bootstrap_servers,  # Kafka broker address
+            'client.id': f"producer_{stream_reader_name}_{task_id}"
+        }
+    )
+    kafka_producer = KafkaProducer(kafka_producer_config)
+    kafka_admin = AdminClient(
+        {'bootstrap.servers': kafka_bootstrap_servers}
+    )
+    producer = Producer(
+        kafka_producer=kafka_producer,
+        kafka_admin=kafka_admin,
+        new_topic_configuration=new_topic_configu,
+    )
+    change_event_handler = ChangeEventHandler(
+        kafka_client=producer,
+        committer_queue=committer_queue,
+        kafka_prefix=kafka_prefix,
+    )
+    application = ProducerFlow(
+        producer_queue=producer_queue,
+        event_handler=change_event_handler,
+        queue_get_timeout=queue_get_timeout,
+    )
+    worker = BaseWorker(
+        manager_pid=manager_pid,
+        manager_create_time=manager_create_time,
+        task_id=task_id,
+        request_queue=request_queue,
+        response_queue=response_queue,
+        queue_get_timeout=queue_get_timeout,
+        queue_put_timeout=queue_put_timeout,
+        application=application,
+    )
+    return worker
+
+
+class ProducerFlowContext(ApplicationContext):
+    build_worker = staticmethod(build_producer_worker)
